Address comments

+++ src/models/message.rs
@@ -2,7 +2,7 @@ use super::{Chat, SuccessfulPayment, MessageEntity, User, Audio, Document, Game,
 
 #[derive(Serialize, Deserialize, Debug, Default)]
 pub struct Message {
-    pub message_id: u32,
+    pub message_id: u64,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub from: Option<User>,
     pub date: u32,
@@ -12,7 +12,7 @@ pub struct Message {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub forward_from_chat: Option<Chat>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub forward_from_message_id: Option<u32>,
+    pub forward_from_message_id: Option<u64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub forward_signature: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -72,9 +72,9 @@ pub struct Message {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub channel_chat_created: Option<bool>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub migrate_to_chat_id: Option<u32>,
+    pub migrate_to_chat_id: Option<u64>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub migrate_from_chat_id: Option<u32>,
+    pub migrate_from_chat_id: Option<u64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub pinned_message: Option<Box<Message>>,
     #[serde(skip_serializing_if = "Option::is_none")]
